Finish wiring reports module to allow for sub-folders for each report
