--- app/report/deansOffice/deansOfficeReport.css
@@ -1,11 +1,0 @@
-.workload-summary__table-label {
-	font-weight: bold;
-	padding-top: 30px;
-	padding-bottom: 5px;
-	padding-left: 20px;
-	font-size: 12px;
-}
-
-.workload-table__last {
-	padding-bottom: 30px;
-}
