+++ app/reports/deansOffice/deansOfficeReportCtrl.html
@@ -1,0 +1,7 @@
+<nav year="{{ year }}" workgroup-id="{{ workgroupId }}"></nav>
+
+<ipa-header page-title="Deans Office Report"></ipa-header>
+
+<div>
+	deans office report
+</div>
