+++ app/reports/reports.html
@@ -27,8 +27,8 @@
 		<![endif]-->
 	</head>
 
-	<body class="page-body ng-cloak" ng-app="workloadSummaryReportApp" ng-cloak ng-controller="SharedCtrl">
-		<base href="/workloadSummaryReport/">
+	<body class="page-body ng-cloak" ng-app="reportsApp" ng-cloak ng-controller="SharedCtrl">
+		<base href="/reports/">
 
 		<div ng-if="loadingView" class="overlay">
 			<spinner class="vertical-center"></spinner>
@@ -52,7 +52,7 @@
 		<script src="/js/sharedApp.js"></script>
 		<script src="/js/clientConfig.js"></script>
 		<script src="/js/snippets.js"></script>
-		<script src="/js/workloadSummaryReportApp.js"></script>
+		<script src="/js/reportsApp.js"></script>
 
 	</body>
 </html>
